update channel get with option --primary
