+++ docs/mkdocs.yml
@@ -522,7 +522,6 @@ nav:
       - channel:
         - channel add: 'cmd/teams/channel/channel-add.md'
         - channel get: 'cmd/teams/channel/channel-get.md'
-        - channel primary get: 'cmd/teams/channel/channel-primary-get.md'
         - channel list: 'cmd/teams/channel/channel-list.md'
         - channel remove: 'cmd/teams/channel/channel-remove.md'
         - channel set: 'cmd/teams/channel/channel-set.md'
