--- src/m365/teams/commands/channel/channel-primary-get.ts
@@ -1,120 +1,0 @@
-import { Logger } from '../../../../cli';
-import { CommandOption } from '../../../../Command';
-import GlobalOptions from '../../../../GlobalOptions';
-import request from '../../../../request';
-import Utils from '../../../../Utils';
-import GraphCommand from '../../../base/GraphCommand';
-import { Channel } from '../../Channel';
-import commands from '../../commands';
-
-interface CommandArgs {
-  options: Options;
-}
-
-interface Options extends GlobalOptions {
-  teamId?: string;
-  teamName?: string;
-}
-
-class TeamsChannelPrimaryGetCommand extends GraphCommand {
-  public get name(): string {
-    return commands.CHANNEL_PRIMARY_GET;
-  }
-
-  public get description(): string {
-    return 'Gets information about the Primary channel for Microsoft Teams team';
-  }
-
-  public getTelemetryProperties(args: CommandArgs): any {
-    const telemetryProps: any = super.getTelemetryProperties(args);
-    telemetryProps.teamId = typeof args.options.teamId !== 'undefined';
-    telemetryProps.teamName = typeof args.options.teamName !== 'undefined';
-    return telemetryProps;
-  }
-
-  private getTeamId(args: CommandArgs): Promise<string> {
-    if (args.options.teamId) {
-      return Promise.resolve(args.options.teamId);
-    }
-
-    const requestOptions: any = {
-      url: `${this.resource}/v1.0/groups?$filter=displayName eq '${encodeURIComponent(args.options.teamName as string)}'`,
-      headers: {
-        accept: 'application/json;odata.metadata=none'
-      },
-      responseType: 'json'
-    };
-
-    return request
-      .get<{ value: [{ id: string, resourceProvisioningOptions: string[] }] }>(requestOptions)
-      .then(response => {
-        const groupItem: { id: string, resourceProvisioningOptions: string[] } | undefined = response.value[0];
-
-        if (!groupItem) {
-          return Promise.reject(`The specified team does not exist in the Microsoft Teams`);
-        }
-
-        if (groupItem.resourceProvisioningOptions.indexOf('Team') === -1) {
-          return Promise.reject(`The specified team does not exist in the Microsoft Teams`);
-        }
-
-        if (response.value.length > 1) {
-          return Promise.reject(`Multiple Microsoft Teams teams with name ${args.options.teamName} found: ${response.value.map(x => x.id)}`);
-        }
-
-        return Promise.resolve(groupItem.id);
-      });
-  }
-
-  public commandAction(logger: Logger, args: CommandArgs, cb: () => void): void {
-    this
-      .getTeamId(args)
-      .then((teamId: string): Promise<Channel> => {
-        const requestOptions: any = {
-          url: `${this.resource}/v1.0/teams/${teamId}/primaryChannel`,
-          headers: {
-            accept: 'application/json;odata.metadata=none'
-          },
-          responseType: 'json'
-        };
-
-        return request
-          .get<Channel>(requestOptions);
-      })
-      .then((res: Channel): void => {
-        logger.log(res);
-        cb();
-      }, (err: any) => this.handleRejectedODataJsonPromise(err, logger, cb));
-  }
-
-  public options(): CommandOption[] {
-    const options: CommandOption[] = [
-      {
-        option: '-i, --teamId [teamId]'
-      },
-      {
-        option: '--teamName [teamName]'
-      }
-    ];
-
-    const parentOptions: CommandOption[] = super.options();
-    return options.concat(parentOptions);
-  }
-
-  public validate(args: CommandArgs): boolean | string {
-    if (args.options.teamId && args.options.teamName) {
-      return 'Specify either teamId or teamName, but not both.';
-    }
-
-    if (!args.options.teamId && !args.options.teamName) {
-      return 'Specify teamId or teamName, one is required';
-    }
-
-    if (args.options.teamId && !Utils.isValidGuid(args.options.teamId as string)) {
-      return `${args.options.teamId} is not a valid GUID`;
-    }
-    return true;
-  }
-}
-
-module.exports = new TeamsChannelPrimaryGetCommand();
